fix(css): add RosettaColor support to css package
